requirements.txt
customtkinter==5.2.0
Pillow==9.2.0
Pillow==10.2.0
pydicom==2.4.4
